+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Release 0.1.2
 
 - introduced ErrWriterNil to catch nil writers
 - better parameter validation
+- better error messages
 - improved tests with configurable broken IO
 - added parallel annotation